Slettet testing XML altinn kode
