+++ gradle.properties
@@ -22,8 +22,6 @@ kafkaVersion=3.8.0
 pdpClientVersion=0.0.6
 dialogportenClientVersion=0.0.2
 swaggerVersion=2.2.28
-jaxbApiVersion =2.4.0-b180830.0359
-syfoXmlCodeGen=2.0.1
 bakgrunnsjobbVersion=1.0.7
 testContainerVersion=1.20.6
 unleashVersion=10.2.2
