--- src/test/kotlin/no/nav/helsearbeidsgiver/sykmelding/altinnFormat/model/Merknad.kt
@@ -1,6 +1,0 @@
-package no.nav.syfo.model
-
-data class Merknad(
-    val type: String,
-    val beskrivelse: String?,
-)
