--- src/test/kotlin/no/nav/helsearbeidsgiver/sykmelding/altinnFormat/model/sykmelding/arbeidsgiver/AktivitetIkkeMuligAGDTO.kt
@@ -1,7 +1,0 @@
-package no.nav.syfo.model.sykmelding.arbeidsgiver
-
-import no.nav.syfo.model.sykmelding.model.ArbeidsrelatertArsakDTO
-
-data class AktivitetIkkeMuligAGDTO(
-    val arbeidsrelatertArsak: ArbeidsrelatertArsakDTO?,
-)
