--- src/test/kotlin/no/nav/helsearbeidsgiver/sykmelding/altinnFormat/model/sykmelding/arbeidsgiver/ArbeidsgiverAGDTO.kt
@@ -1,6 +1,0 @@
-package no.nav.syfo.model.sykmelding.arbeidsgiver
-
-data class ArbeidsgiverAGDTO(
-    val navn: String?,
-    val yrkesbetegnelse: String?,
-)
