--- src/test/kotlin/no/nav/helsearbeidsgiver/sykmelding/altinnFormat/model/sykmelding/arbeidsgiver/KontaktMedPasientAGDTO.kt
@@ -1,7 +1,0 @@
-package no.nav.syfo.model.sykmelding.arbeidsgiver
-
-import java.time.LocalDate
-
-data class KontaktMedPasientAGDTO(
-    val kontaktDato: LocalDate?,
-)
