--- src/test/kotlin/no/nav/helsearbeidsgiver/sykmelding/altinnFormat/model/sykmelding/arbeidsgiver/PrognoseAGDTO.kt
@@ -1,6 +1,0 @@
-package no.nav.syfo.model.sykmelding.arbeidsgiver
-
-data class PrognoseAGDTO(
-    val arbeidsforEtterPeriode: Boolean,
-    val hensynArbeidsplassen: String?,
-)
