--- src/test/kotlin/no/nav/helsearbeidsgiver/sykmelding/altinnFormat/model/sykmelding/arbeidsgiver/SykmeldingsperiodeAGDTO.kt
@@ -1,16 +1,0 @@
-package no.nav.syfo.model.sykmelding.arbeidsgiver
-
-import no.nav.syfo.model.sykmelding.model.GradertDTO
-import no.nav.syfo.model.sykmelding.model.PeriodetypeDTO
-import java.time.LocalDate
-
-data class SykmeldingsperiodeAGDTO(
-    val fom: LocalDate,
-    val tom: LocalDate,
-    val gradert: GradertDTO?,
-    val behandlingsdager: Int?,
-    val innspillTilArbeidsgiver: String?,
-    val type: PeriodetypeDTO,
-    val aktivitetIkkeMulig: AktivitetIkkeMuligAGDTO?,
-    val reisetilskudd: Boolean,
-)
