--- src/test/kotlin/no/nav/helsearbeidsgiver/sykmelding/altinnFormat/model/sykmelding/arbeidsgiver/UtenlandskSykmeldingAGDTO.kt
@@ -1,5 +1,0 @@
-package no.nav.syfo.model.sykmelding.arbeidsgiver
-
-data class UtenlandskSykmeldingAGDTO(
-    val land: String,
-)
