--- src/test/kotlin/no/nav/helsearbeidsgiver/sykmelding/altinnFormat/model/sykmelding/model/AdresseDTO.kt
@@ -1,9 +1,0 @@
-package no.nav.syfo.model.sykmelding.model
-
-data class AdresseDTO(
-    val gate: String?,
-    val postnummer: Int?,
-    val kommune: String?,
-    val postboks: String?,
-    val land: String?,
-)
