--- src/test/kotlin/no/nav/helsearbeidsgiver/sykmelding/altinnFormat/model/sykmelding/model/ArbeidsrelatertArsakDTO.kt
@@ -1,6 +1,0 @@
-package no.nav.syfo.model.sykmelding.model
-
-class ArbeidsrelatertArsakDTO(
-    val beskrivelse: String?,
-    val arsak: List<ArbeidsrelatertArsakTypeDTO>,
-)
