--- src/test/kotlin/no/nav/helsearbeidsgiver/sykmelding/altinnFormat/model/sykmelding/model/GradertDTO.kt
@@ -1,6 +1,0 @@
-package no.nav.syfo.model.sykmelding.model
-
-data class GradertDTO(
-    val grad: Int,
-    val reisetilskudd: Boolean,
-)
