--- src/test/kotlin/no/nav/helsearbeidsgiver/sykmelding/altinnFormat/model/sykmelding/model/PeriodetypeDTO.kt
@@ -1,9 +1,0 @@
-package no.nav.syfo.model.sykmelding.model
-
-enum class PeriodetypeDTO {
-    AKTIVITET_IKKE_MULIG,
-    AVVENTENDE,
-    BEHANDLINGSDAGER,
-    GRADERT,
-    REISETILSKUDD,
-}
