--- src/test/kotlin/no/nav/helsearbeidsgiver/sykmelding/altinnFormat/model/sykmelding/model/TidligereArbeidsgiverDTO.kt
@@ -1,7 +1,0 @@
-package no.nav.syfo.model.sykmelding.model
-
-data class TidligereArbeidsgiverDTO(
-    val orgNavn: String,
-    val orgnummer: String,
-    val sykmeldingsId: String,
-)
