--- src/test/kotlin/no/nav/helsearbeidsgiver/sykmelding/altinnFormat/model/sykmeldingstatus/KafkaMetadataDTO.kt
@@ -1,10 +1,0 @@
-package no.nav.syfo.model.sykmeldingstatus
-
-import java.time.OffsetDateTime
-
-data class KafkaMetadataDTO(
-    val sykmeldingId: String,
-    val timestamp: OffsetDateTime,
-    val fnr: String,
-    val source: String,
-)
