--- src/test/kotlin/no/nav/helsearbeidsgiver/sykmelding/altinnFormat/model/sykmeldingstatus/SykmeldingStatusKafkaEventDTO.kt
@@ -1,14 +1,0 @@
-package no.nav.syfo.model.sykmeldingstatus
-
-import no.nav.syfo.model.sykmelding.model.TidligereArbeidsgiverDTO
-import java.time.OffsetDateTime
-
-data class SykmeldingStatusKafkaEventDTO(
-    val sykmeldingId: String,
-    val timestamp: OffsetDateTime,
-    val statusEvent: String,
-    val arbeidsgiver: ArbeidsgiverStatusDTO? = null,
-    val sporsmals: List<SporsmalOgSvarDTO>? = null,
-    val erSvarOppdatering: Boolean? = null,
-    val tidligereArbeidsgiver: TidligereArbeidsgiverDTO? = null,
-)
